Authentication added in all project endpoints

+++ src/controllers/ProjectController.js
@@ -6,6 +6,13 @@ module.exports = {
     try {
       const { client_id } = req.params;
       const { title, description, value, deadline } = req.body;
+
+      const client = await Client.findOne({ where: { id: client_id, user_id: req.user.id } });
+
+      if (!client) {
+        return res.status(404).send({ error: 'Client not found!' });
+      }
+
       const project = await Project.create({
         client_id,
         title,
@@ -29,7 +36,7 @@ module.exports = {
         include: { association: 'projects' },
       });
 
-      if (!client) {
+      if (!client || client.user_id !== req.user.id) {
         return res.status(404).send({ error: 'The client doesn\'t exist.' });
       }
 
@@ -43,6 +50,11 @@ module.exports = {
   async updateProject(req, res) {
     try {
       const { client_id, project_id } = req.params;
+      const client = await Client.findOne({ where: { id: client_id, user_id: req.user.id } });
+
+      if (!client) {
+        return res.status(404).send({ error: 'Client not found!' });
+      }
 
       const project = await Project.update(req.body, {
         where: { id: project_id, client_id },
@@ -63,7 +75,17 @@ module.exports = {
     try {
       const { client_id, project_id } = req.params;
 
+      const client = await Client.findOne({ where: { id: client_id, user_id: req.user.id } });
+
+      if (!client) {
+        return res.status(404).send({ error: 'Client not found!' });
+      }
+
       const project = await Project.destroy({ where: { id: project_id, client_id } });
+
+      if (!project) {
+        return res.status(404).send({ error: 'Project not found! ' });
+      }
 
       res.json(project);
     }
